Improve the tests on the IMo
There was an ambiguity in the expression “mock IMo”, as it could refer
to the PTEP IMo or to the IMo files in the test/ directory. Now the IMo
in the `test/` directory is called `test_imo`, and the directory
containing the YAML file has been renamed accordingly.
